# Add `inrangecount` for efficient counting of neighbors