--- a/src/kd_tree.jl
+++ b/src/kd_tree.jl
@@ -106,8 +106,8 @@ function KDTree(data::Vector{T},
                 storedata::Bool = true,
                 reorder::Bool = true,
                 reorderbuffer::Vector{T} = Vector{T}(undef, 0, 0)) where {T <: AbstractFloat, M <: MinkowskiMetric}
-   KDTree(hcat(data), metric, leafsize = leafsize, storedata = storedata, reorder = reorder,
-           reorderbuffer = reorderbuffer_points)
+   KDTree(reshape(data, length(data), 1), metric, leafsize = leafsize, storedata = storedata,
+          reorder = reorder, reorderbuffer = reorderbuffer_points)
 end
 
 function build_KDTree(index::Int,