--- a/src/tree_ops.jl
+++ b/src/tree_ops.jl
@@ -114,29 +114,34 @@ end
 # stop computing the distance function as soon as we reach the desired radius.
 # This will probably prevent SIMD and other optimizations so some care is needed
 # to evaluate if it is worth it.
-@inline function add_points_inrange!(idx_in_ball::AbstractVector{Int}, tree::NNTree,
+@inline function add_points_inrange!(idx_in_ball::Union{Nothing, AbstractVector{Int}}, tree::NNTree,
                                      index::Int, point::AbstractVector, r::Number, do_end::Bool)
+    count = 0
     for z in get_leaf_range(tree.tree_data, index)
         idx = tree.reordered ? z : tree.indices[z]
         dist_d = evaluate(tree.metric, tree.data[idx], point, do_end)
         if dist_d <= r
-            push!(idx_in_ball, idx)
+            count += 1
+            idx_in_ball !== nothing && push!(idx_in_ball, idx)
         end
     end
+    return count
 end
 
 # Add all points in this subtree since we have determined
 # they are all within the desired range
-function addall(tree::NNTree, index::Int, idx_in_ball::Vector{Int})
+function addall(tree::NNTree, index::Int, idx_in_ball::Union{Nothing, Vector{Int}})
     tree_data = tree.tree_data
-    if isleaf(tree.tree_data.n_internal_nodes, index)
-        for z in get_leaf_range(tree.tree_data, index)
+    count = 0
+    if isleaf(tree_data.n_internal_nodes, index)
+        for z in get_leaf_range(tree_data, index)
             idx = tree.reordered ? z : tree.indices[z]
-            push!(idx_in_ball, idx)
+            count += 1
+            idx_in_ball !== nothing && push!(idx_in_ball, idx)
         end
-        return
     else
-        addall(tree, getleft(index), idx_in_ball)
-        addall(tree, getright(index), idx_in_ball)
+        count += addall(tree, getleft(index), idx_in_ball)
+        count += addall(tree, getright(index), idx_in_ball)
     end
+    return count
 end