--- a/test/test_inrange.jl
+++ b/test/test_inrange.jl
@@ -8,12 +8,18 @@
 
                 idxs = inrange(tree, [1.1, 1.1, 1.1], 0.2, dosort)
                 @test idxs == [8] # Only corner 8 at least 0.2 distance away from [1.1, 1.1, 1.1]
+                counts = inrangecount(tree, [1.1, 1.1, 1.1], 0.2)
+                @test counts == 1
 
                 idxs = inrange(tree, [0.0, 0.0, 0.5], 0.6, dosort)
                 @test idxs == [1, 2] # Corner 1 and 2 at least 0.6 distance away from [0.0, 0.0, 0.5]
+                counts = inrangecount(tree, [0.0, 0.0, 0.5], 0.6)
+                @test counts == 2
 
                 idxs = inrange(tree, [0, 0, 0], 0.6, dosort)
                 @test idxs == [1]
+                counts = inrangecount(tree, [0, 0, 0], 0.6)
+                @test counts == 1
 
                 X = [0.0 0.0; 0.0 0.0; 0.5 0.0]
                 idxs1 = inrange(tree, X, 0.6, dosort)
@@ -21,19 +27,31 @@
                 @test idxs1 == idxs2
                 @test idxs1[1] == [1,2]
                 @test idxs1[2] == [1]
+                counts1 = inrangecount(tree, X, 0.6)
+                counts2 = inrangecount(tree, view(X,:,1:2), 0.6)
+                @test counts1 == counts2
+                @test counts1 == [2, 1]
 
                 idxs = inrange(tree, [SVector{3,Float64}(0.0, 0.0, 0.5), SVector{3,Float64}(0.0, 0.0, 0.0)], 0.6, dosort)
                 @test idxs[1] == [1,2]
                 @test idxs[2] == [1]
+                counts = inrangecount(tree, [SVector{3,Float64}(0.0, 0.0, 0.5), SVector{3,Float64}(0.0, 0.0, 0.0)], 0.6)
+                @test counts == [2, 1]
 
                 idxs = inrange(tree, [0.33333333333, 0.33333333333, 0.33333333333], 1, dosort)
                 @test idxs == [1, 2, 3, 5]
+                counts = inrangecount(tree, [0.33333333333, 0.33333333333, 0.33333333333], 1)
+                @test counts == 4
 
                 idxs = inrange(tree, [0.5, 0.5, 0.5], 0.2, dosort)
                 @test idxs == []
+                counts = inrangecount(tree, [0.5, 0.5, 0.5], 0.2)
+                @test counts == 0
 
                 idxs = inrange(tree, [0.5, 0.5, 0.5], 1.0, dosort)
                 @test idxs == [1, 2, 3, 4, 5, 6, 7, 8]
+                counts = inrangecount(tree, [0.5, 0.5, 0.5], 1.0)
+                @test counts == 8
 
                 @test_throws ArgumentError inrange(tree, rand(3), -0.1)
                 @test_throws ArgumentError inrange(tree, rand(5), 1.0)
@@ -41,10 +59,14 @@
                 empty_tree = TreeType(rand(3,0), metric)
                 idxs = inrange(empty_tree, [0.5, 0.5, 0.5], 1.0)
                 @test idxs == []
+                counts = inrangecount(empty_tree, [0.5, 0.5, 0.5], 1.0)
+                @test counts == 0
 
                 one_point_tree = TreeType([0.5, 0.5, 0.5], metric)
                 idxs = inrange(one_point_tree, data, 1.0)
                 @test idxs == repeat([[1]], size(data, 2))
+                counts = inrangecount(one_point_tree, data, 1.0)
+                @test counts == repeat([1], size(data, 2))
             end
             data = [0.0 0.0 0.0 0.0 1.0 1.0 1.0 1.0;
                     0.0 0.0 1.0 1.0 0.0 0.0 1.0 1.0;