Added an efficient method for counting neighbors without storing their indexes.

**EDIT: This is out of date, see post below**

This was primarily achieved by reusing the index-accumulating vector for `inrange_points!`

```
@btime [(inrangecount(ctrees[j], (@view c.coordinates[:, c.abovethreshold]), r) .- k) ./ r ^ 2 for r ∈ radiussteps]
  72.972 ms (701 allocations: 1.76 MiB)

@btime [(length.(inrange(ctrees[j], (@view c.coordinates[:, c.abovethreshold]), r)) .- k) ./ r ^ 2 for r ∈ radiussteps]
  190.149 ms (30129 allocations: 163.86 MiB)
```

The above calculates a density gradient around each point, with 100 test points and 25 radius steps against a tree of ~20,000.

This PR cuts execution time of the above code by 2.6x, allocations by 28.7x, and allocation amount by 93.1x.

An important feature of this PR is it scales well to more test points. This implementation allowed ~20,000 points to be processed against ~20,000 points in 12 s on my machine. The previous version took 3 min. And I need it to work on hundreds of thousands or millions of points.

I'll add tests and documentation if you agree with it.