@KristofferC 2021-09-01
I think it probably makes sense to have this here.

The speedup here is purely from not allocating the output data structures? There is no "algothmic improvement", right?