@BioTurboNick 2021-09-03
Because I can't leave well-enough alone, I made one last change, which is to have the `inrange` kernel methods count and return an integer instead of `nothing`, and if `nothing` is passed for an index accumulator, index accumulating doesn't occur.

New performance:
```
Before: 16.318 s (701 allocations: 19.67 MiB)
After: 2.195 s (504 allocations: 16.23 MiB)
```

And to ensure the index-accumulating method is still okay:
```
Before: 761.788 ms (115011 allocations: 770.46 MiB)
After: 746.208 ms (115011 allocations: 770.46 MiB)
```