@KristofferC 2022-03-23
Looks good but I wonder if `countinrange` is a better name? But it looks a bit weird, maybe `count_inrange`? Any thoughts?