@BioTurboNick 2022-03-23
Thanks! I could go either way.

I think `inrangecount`/`inrange_count` would benefit from sharing the prefix, for discoverability.

But I see precedence for `countx` in Base, so that seems like the convention to use.

I'll go with `count_inrange`.