@KristofferC 2022-03-23
> I think inrangecount/inrange_count would benefit from sharing the prefix, for discoverability.

That's a good point. I think that convinces me. We can go back to the earlier version. Sorry for the nitpicking. Should be easy enough to just drop the new commits?